requirements.txt
pycoingecko==1.4.0
python-binance==0.7.9
python-decouple==3.4
requests==2.25.1